Added AnimationComponent. Added Scene::register_system(). Texture Description refactoring. GraphicsDevice added set_sampler_state. Hidden window optimizition.

+++ Gengine/main.cpp
@@ -112,7 +112,7 @@ int WINAPI wWinMain(
         auto now = std::chrono::high_resolution_clock::now();
         auto duration = std::chrono::duration_cast<std::chrono::milliseconds>(now - time);
 
-        form->main_scene->world()->tick(1.f/duration.count());
+        form->main_scene->world()->tick(1.f/(float)duration.count());
 		
         time = now;
 	}
